pyproject.toml
[project]
name = "alfred"
version = "0.2.0"
description = "Alfred — personal assistant bot with skill registry"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "mcp>=1.5",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "python-telegram-bot>=21.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
